Completed making a simple backend

+++ api/routes/users.js
@@ -5,44 +5,69 @@ const bcrypt = require("bcrypt");
 
 // update
 
-router.put("/:id", async(req,res)=>{
-    if (req.body.userId === req.params.id){
-        if (req.body.password){
+router.put("/:id", async (req, res) => {
+    if (req.body.userId === req.params.id) {
+        if (req.body.password) {
             const salt = await bcrypt.genSalt(10);
             req.body.password = await bcrypt.hash(req.body.password, salt);
         }
-        try{
-            const updateUser = await User.findByIdAndUpdate(req.params.id,{
+        try {
+            const updateUser = await User.findByIdAndUpdate(req.params.id, {
                 $set: req.body,
-            },{new:true});
+            }, { new: true });
             res.status(200).json(updateUser);
         }
-        catch(err){
+        catch (err) {
             res.status(500).json(err)
         }
-    }else{
-        res.status(401).json("This is not your account")        
+    } else {
+        res.status(401).json("This is not your account")
     }
 })
 
 // delete
-router.delete("/:id", async(req,res)=>{
-    if (req.body.userId === req.params.id){
+router.delete("/:id", async (req, res) => {
+    if (req.body.userId === req.params.id) {
         try {
-            const user = await User.findById(req.params.id);
-            try{
-                await Post.deleteMany({userName: user.userName})
+            const user = await User.findById(req.params.id); 
+            
+            if(!user){
+                res.status(404).json("User not found");
+                return;
+            }
+            try {
+                await Post.deleteMany({ userName: user.userName })
                 await User.findByIdAndDelete(req.params.id)
                 res.status(200).json("User has been deleted");
             }
-            catch(err){
+            catch (err) {
                 res.status(500).json(err)
-            }            
+            }
         } catch (error) {
             res.status(404).json("User not found")
         }
-    }else{
-        res.status(401).json("This is not your account")        
+    } else {
+        res.status(401).json("This is not your account")
+    }
+})
+
+// Get user
+
+router.get("/:id", async (req, res) => {
+    try {
+
+        const user = await User.findById(req.params.id);
+        
+        if(!user){
+            res.status(404).json("User not found");
+            return;
+        }
+        
+        const { password, ...others } = user._doc;
+        
+        res.status(200).json(others);
+    } catch (err) {
+        res.status(500).json(err);
     }
 })
